+++ common/model/src/main/java/module-info.java
@@ -25,7 +25,6 @@
  * <li>{@link is.codion.common.model.selection.MultiSelection}
  * <li>{@link is.codion.common.model.condition.ConditionModel}
  * <li>{@link is.codion.common.model.condition.TableConditionModel}
- * <li>{@link is.codion.common.model.preferences.FilePreferencesFactory}
  * <li>{@link is.codion.common.model.preferences.UserPreferences}
  * <li>{@link is.codion.common.model.summary.SummaryModel}
  * <li>{@link is.codion.common.model.summary.TableSummaryModel}