+++ framework/model/src/main/java/is/codion/framework/model/package-info.java
@@ -21,6 +21,9 @@
  * <ul>
  * <li>{@link is.codion.framework.model.EntityApplicationModel#SAVE_DEFAULT_USERNAME}
  * <li>{@link is.codion.framework.model.EntityApplicationModel#USER}
+ * <li>{@link is.codion.framework.model.EntityApplicationModel#USER_PREFERENCES}
+ * <li>{@link is.codion.framework.model.EntityApplicationModel#RESTORE_DEFAULT_PREFERENCES}
+ * <li>{@link is.codion.framework.model.EntityApplicationModel#PREFERENCES_KEY}
  * <li>{@link is.codion.framework.model.EntityEditModel#EDIT_EVENTS}
  * <li>{@link is.codion.framework.model.EntityEditModel.EntityEditor#PERSIST_FOREIGN_KEYS}
  * <li>{@link is.codion.framework.model.EntityQueryModel#LIMIT}