+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -23,8 +23,6 @@
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#CONFIRM_EXIT}
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#SHOW_STARTUP_DIALOG}
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#CACHE_ENTITY_PANELS}
- * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#USER_PREFERENCES}
- * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#RESTORE_DEFAULT_PREFERENCES}
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#CALL_SYSTEM_EXIT}
  * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#VALID_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#MODIFIED_INDICATOR}